+++ src/C2nRthsController.cs
@@ -124,7 +124,7 @@ namespace PDT.Plugins.Crestron.IO
             if (cresnetBridge != null)
             {
                 Debug.Console(0, "Device {0} is a valid cresnet master - creating new C2nRths", parentKey);
-                return new C2nRths(cresnetId, cresnetBridge.CresnetBranches[branchId]);
+                return new C2nRths(cresnetId, cresnetBridge.CresnetBranches[(uint)branchId]);
             }
             Debug.Console(0, "Device {0} is not a valid cresnet master", parentKey);
             return null;
