+++ src/GlsOirOccupancySensorController.cs
@@ -71,7 +71,7 @@ namespace PDT.Plugins.Crestron.IO
             if (cresnetBridge != null)
             {
                 Debug.Console(0, "Device {0} is a valid cresnet master - creating new GlsOirCCn", parentKey);
-                return new GlsOirCCn(cresnetId, cresnetBridge.CresnetBranches[branchId]);
+                return new GlsOirCCn(cresnetId, cresnetBridge.CresnetBranches[(uint)branchId]);
             }
             Debug.Console(0, "Device {0} is not a valid cresnet master", parentKey);
             return null;
