+++ src/StatusSignController.cs
@@ -195,7 +195,7 @@ namespace PDT.Plugins.Crestron.IO
             if (cresnetBridge != null)
             {
                 Debug.Console(0, "Device {0} is a valid cresnet master - creating new StatusSign", parentKey);
-                return new StatusSign(cresnetId, cresnetBridge.CresnetBranches[branchId]);
+                return new StatusSign(cresnetId, cresnetBridge.CresnetBranches[(uint)branchId]);
             }
             Debug.Console(0, "Device {0} is not a valid cresnet master", parentKey);
             return null;
